backend/requirements.txt
flask
flask-cors
scikit-learn
pandas
numpy
joblib
pydantic
python-dotenv
PyJWT
shap
gunicorn
